+++ engines/ring/graphics/aquator.h
@@ -70,10 +70,10 @@ public:
 	static void set(int *pixel, int val1, int val2, int val3);
 	static void set(PixelTriplet *from, PixelData *pixel, float a1, float a2, float a3);
 	static void set(PixelTriplet *from, PixelData *pixel, PixelData *pixel2);
-	static void add(int *pixel, int val);
+	static void add(float *pixel, float val);
 	static void add(PixelData *triplet, PixelData *from, float a1, float a2, float a3);
 	static void substract(int *pixel, int val);
-	static void multiply(int *pixel, double val);
+	static void multiply(float *pixel, float val);
 	static void multiply(PixelData *from, PixelData *pixel, float val);
 	static void divide(PixelData *pixel);
 };
